Implemented calendar for Date/Time selection
Still need to create handler for when a time slot is selected
